+++ arch/ppc/arch_callbacks.h
@@ -5,8 +5,8 @@
 
 uint32_t tlib_read_tbl(void);
 uint32_t tlib_read_tbu(void);
-uint32_t tlib_read_decrementer(void);
+uint64_t tlib_read_decrementer(void);
 uint32_t tlib_is_vle_enabled(void);
-void tlib_write_decrementer(uint32_t value);
+void tlib_write_decrementer(uint64_t value);
 
 #endif
